refactor: move endpoints of view to view controller
