+++ src/app.module.ts
@@ -5,20 +5,13 @@ import { GraphQLModule } from '@nestjs/graphql';
 import { AppController } from './app.controller';
 import { AuthModule } from './auth/auth.module';
 import { AuthorModule } from './author/author.module';
-import { AuthorService } from './author/author.service';
 import { BookModule } from './book/book.module';
-import { BookService } from './book/book.service';
 import { CategoryModule } from './category/category.module';
-import { CategoryService } from './category/category.service';
 import { OrderModule } from './order/order.module';
-import { OrderService } from './order/order.service';
 import { PrismaModule } from './prisma/prisma.module';
 import { PromotionListModule } from './promotion-list/promotion-list.module';
-import { PromotionListService } from './promotion-list/promotion-list.service';
 import { RatingReviewModule } from './rating-review/rating-review.module';
-import { RatingReviewService } from './rating-review/rating-review.service';
 import { UserModule } from './user/user.module';
-import { UserService } from './user/user.service';
 
 @Module({
   imports: [
@@ -42,14 +35,5 @@ import { UserService } from './user/user.service';
     UserModule,
   ],
   controllers: [AppController],
-  providers: [
-    AuthorService,
-    CategoryService,
-    PromotionListService,
-    OrderService,
-    RatingReviewService,
-    UserService,
-    BookService,
-  ],
 })
 export class AppModule {}
