+++ src/category/category.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
+import { CategoryViewController } from './category-view.controller';
 import { CategoryController } from './category.controller';
 import { CategoryService } from './category.service';
 
 @Module({
-  controllers: [CategoryController],
-  providers: [CategoryService]
+  controllers: [CategoryController, CategoryViewController],
+  providers: [CategoryService],
 })
 export class CategoryModule {}
