+++ src/rating-review/rating-review.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
+import { RatingReviewViewController } from './rating-review-view.controller';
 import { RatingReviewController } from './rating-review.controller';
 import { RatingReviewService } from './rating-review.service';
 
 @Module({
-  controllers: [RatingReviewController],
-  providers: [RatingReviewService]
+  controllers: [RatingReviewController, RatingReviewViewController],
+  providers: [RatingReviewService],
 })
 export class RatingReviewModule {}
